requirements.txt
cachetools==4.2.2
Pillow==8.4.0
